Move CSS files to same directory as target components
