--- src/Styles/CheckPage.css
@@ -1,27 +1,0 @@
-.header-div {
-  background-color: orange;
-  height: 130px;
-  display: flex;
-  position: fixed;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: space-between;
-}
-
-.back-button {
-  position: absolute;
-  width: 178px;
-  height: 70px;
-  left: 25px;
-  top: 48px;
-
-  background: #154734;
-  box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
-  border-radius: 20px;
-}
-
-.title {
-  font-size: 50px;
-  color: white;
-  vertical-align: middle;
-}
